LPS-69997 Move to better package
